Don't build nvim every time

+++ setup.sh
@@ -1,15 +1,19 @@
 #! /bin/bash
 
-sudo apt-get install vim fish pandoc ccls npm fzf curl fd-find
+sudo apt-get install vim fish pandoc ccls npm fzf curl fd-find -y
 
 # install neovim
-sudo apt-get install ninja-build gettext libtool libtool-bin autoconf automake cmake g++ pkg-config unzip curl doxygen
-git clone https://github.com/neovim/neovim
-cd neovim && make CMAKE_BUILD_TYPE=RelWithDebInfo
-git checkout stable
-sudo make install
-cd ..
-rm neovim -rf
+echo "Would you like to reinstall Neovim from source? (Y/n)"
+read rebuildNvim
+if ["$rebuildNvim" == "Y"] || ["$rebuildNvim" == "y"];then
+	sudo apt-get install ninja-build gettext libtool libtool-bin autoconf automake cmake g++ pkg-config unzip curl doxygen -y
+	git clone https://github.com/neovim/neovim
+	cd neovim && make CMAKE_BUILD_TYPE=RelWithDebInfo
+	git checkout stable
+	sudo make install
+	cd ..
+	rm neovim -rf
+fi
 
 ln -s /run/media/$USER ~/Drives
